requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
coverage==4.5.1
flake8===3.5.0
flask-debugtoolbar==0.10.1
Flask-JWT-Extended==3.13.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
psycopg2==2.7.4
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-RESTful==0.3.6
passlib==1.7.1
PyJWT==1.6.4
PyQRCode==1.2.1
psycopg2-binary==2.8.4